Move optimistic locking specs into shared specs for persister, queries
resolves #457
