+++ spec/valkyrie/persistence/postgres/persister_spec.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 require 'spec_helper'
 require 'valkyrie/specs/shared_specs'
-require 'valkyrie/specs/shared_specs/locking_persister'
 
 RSpec.describe Valkyrie::Persistence::Postgres::Persister do
   let(:query_service) { adapter.query_service }
@@ -9,7 +8,6 @@ RSpec.describe Valkyrie::Persistence::Postgres::Persister do
 
   let(:persister) { adapter.persister }
   it_behaves_like "a Valkyrie::Persister"
-  it_behaves_like "a Valkyrie locking persister"
 
   context "single value behavior" do
     before do
