+++ spec/valkyrie/persistence/solr/persister_spec.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 require 'spec_helper'
 require 'valkyrie/specs/shared_specs'
-require 'valkyrie/specs/shared_specs/locking_persister'
 
 RSpec.describe Valkyrie::Persistence::Solr::Persister do
   let(:query_service) { adapter.query_service }
@@ -9,7 +8,6 @@ RSpec.describe Valkyrie::Persistence::Solr::Persister do
   let(:adapter) { Valkyrie::Persistence::Solr::MetadataAdapter.new(connection: client) }
   let(:client) { RSolr.connect(url: SOLR_TEST_URL) }
   it_behaves_like "a Valkyrie::Persister"
-  it_behaves_like "a Valkyrie locking persister"
 
   context "when given additional persisters" do
     let(:adapter) { Valkyrie::Persistence::Solr::MetadataAdapter.new(connection: client, resource_indexer: indexer) }
